pyproject.toml
[project]
name = "pixflow"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "ELGINA",email = "anstasiaelgina738@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
